Extracted TokenInjectingHTTPClient
